// src/types/store/action.types.ts
export enum AuthActionTypes {
  CHECK_AUTH = "CHECK_AUTH",
  LOGIN = "LOGIN",
  REFRESH = "REFRESH",
  LOGOUT = "LOGOUT",
  FETCH_LOGO = "FETCH_LOGO",
  CHECK_ACCESS = "CHECK_ACCESS",
  GOOGLE_LOGIN = "GOOGLE_LOGIN",
}

export enum PropertyActionTypes {
  FETCH_PROPERTY_CLASSIFICATIONS = "FETCH_PROPERTY_CLASSIFICATIONS",
  FETCH_PROPERTY_TYPES = "FETCH_PROPERTY_TYPES",
  FETCH_PROPERTY_FUNDS = "FETCH_PROPERTY_FUNDS",
  FETCH_PROPERTY_SELLINGBANKS = "FETCH_PROPERTY_SELLINGBANKS",
  FETCH_COUNTRIES = "FETCH_COUNTRIES",
  FETCH_STATES = "FETCH_STATES",
  FETCH_CITIES = "FETCH_CITIES",
  FETCH_STATUS = "FETCH_STATUS",
  FETCH_AMENITIES = "FETCH_AMENITIES",
  FETCH_AMENITIES2 = "FETCH_AMENITIES2",
}

export enum InvestorActionTypes {
  FETCH_INVESTOR_PROFILE = "FETCH_INVESTOR_PROFILE",
  FETCH_INVESTOR_FUNDS = "FETCH_INVESTOR_FUNDS",
}